requirements.txt
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
bzip2==1.0.8
ca-certificates==2025.1.31
click==8.1.8
comm==0.2.2
cyrus-sasl==2.1.27
debugpy==1.8.12
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
flask==3.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
icu==75.1
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
krb5==1.21.3
libcxx==19.1.7
libedit==3.1.20240808
libexpat==2.6.4
libffi==3.4.2
liblzma==5.6.3
liblzma-devel==5.6.3
libntlm==1.8
libpq==17.2
libsodium==1.0.20
libsqlite==3.48.0
libzlib==1.3.1
markupsafe==3.0.2
matplotlib-inline==0.1.7
ncurses==6.5
nest-asyncio==1.6.0
openldap==2.6.9
openssl==3.4.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.3.1
platformdirs==4.3.6
prompt-toolkit==3.0.50
psutil==6.1.1
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pygments==2.19.1
python==3.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python_abi==3.11
pyzmq==26.2.1
readline==8.2
setuptools==75.8.0
six==1.17.0
sqlalchemy==2.0.38
sqlite==3.48.0
stack_data==0.6.3
tk==8.6.13
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
typing_extensions==4.12.2
tzdata==2025a
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
xz==5.6.3
xz-gpl-tools==5.6.3
xz-tools==5.6.3
zeromq==4.3.5
zipp==3.21.0
